requirements.txt
numpy                     1.26.4          
matplotlib                3.8.0           
pandas                    2.1.4           
opencv-python             4.9.0.80                 
opencv-python-headless    4.11.0.86                
scikit-multilearn         0.2.0                    
pillow                    11.1.0         
scipy                     1.11.0                   
pytorch                   2.2.0           
torchaudio                2.2.0                    
torchvision               0.17.0                   
tensorboard               2.19.0                   
rich                      13.9.4          
timm                      1.0.19                   
einops                    0.8.1                    
openpyxl                  3.1.5                    
ml-collections            1.1.0                    
